% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_parametric_plot.R
\name{better_parametric_plot}
\alias{better_parametric_plot}
\title{Plot better plots of parametric predictors in QGAMs}
\usage{
better_parametric_plot(
  qgam,
  pred,
  print.summary,
  plot.old,
  size,
  color,
  alpha,
  order)
}
\arguments{
\item{qgam}{A qgam object, created with \code{qgam} or extracted from a \code{mqgam} object.}

\item{pred}{A named list of the levels to use for the predictor terms to plot (same as those specified for \code{plot_parametric}).}

\item{print.summary}{Logical: whether or not to print summary.}

\item{plot.old}{Plot the original \code{plot_parametric} as well.}

\item{order}{Specify the order with which the levels given in \code{pred} should be plotted.}

\item{size}{Size argument for the ggplot object; specifies the size of points and lines.}

\item{color}{Color argument for the ggplot object; specifies the color of points and lines.}

\item{alpha}{Alpha argument for the ggplot object; specifies the transparency of points and lines.}
}
\value{
A ggplot object.
}
\description{
\code{better_parametric_plot} creates a point plot with confidence interval ranges using \code{ggplot2}. It basically is a wrapper for \code{itsadug}'s
\code{plot_parametric} function.
}
\examples{
better_parametric_plot(qgam = tmp.x.1,
  pred = list(Condition = c("matched", "mismatched")),
  size = 2,
  color = c("red", "green"),
  alpha = 0.2,
  order = c("mismatched", "matched"))

better_parametric_plot(qgam = tmp.x.1,
  pred = list(Condition = c("matched", "mismatched"))) +
  theme_classic() +
  labs(subtitle = "You can basically add all ggplot functions and arguments you are familiar with.")

}
\references{
van Rij J, Wieling M, Baayen R, & van Rijn H (2020). itsadug: Interpreting Time Series and Autocorrelated Data Using GAMMs. R package version 2.4.

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York.
}
\author{
D. Schmitz
}
