% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xyt.R
\name{extract_xyt}
\alias{extract_xyt}
\title{Extract x, y, and t coordinates of time-normalized mouse-tracking data}
\usage{
extract_xyt(
  tn_data,
  ID_column,
  timestamps,
  verbose = TRUE)
}
\arguments{
\item{mousetrap}{A \code{mousetrap} object containing time-normalized data (commonly achieved using mousetrap::mt_time_normalize).}

\item{tn_data}{\emph{obsolete}: Specify \code{mousetrap} instead. For now, \code{tn_data} remains useable. It will be depricated in due time. \code{tn_data} is the time-normalized mouse-tracking data. Can be extracted from the \code{mousetrap} object via \code{df$tn_trajectories}.}

\item{ID_column}{To maintain identification of individual trials, specify your ID variable. In older versions, this had to be part of \code{tn_data}.}

\item{timestamps}{Specify the number of timestamps you have used for time-normalizing.}

\item{verbose}{If \code{TRUE} (which is the default), a progress bar is displayed.}
}
\value{
A data frame.
\itemize{
\item \code{x_key} - Numbered x values per mouse-track ID.
\item \code{x_value} - The x coordinate value.
\item \code{y_key} - Numbered y values per mouse-track ID.
\item \code{y_value} - The y coordinate value.
\item \code{t_key} - Numbered t values per mouse-track ID.
\item \code{t_value} - The time value.
\item \code{ID} - Mouse-track IDs.
}
}
\description{
\code{extract_xyt} extracts the x, y, and t coordinates, i.e. time values, from a time-normalized \code{mousetrap} object. It basically is a wrap of the
\code{extract_x}, \code{extract_y}, and \code{extract_t} functions.
}
\examples{
xyt_data <- extract_xyt(mousetrap = mtqgam_mousetrap,
  ID_column = "ID",
  timestamps = 100,
  verbose = FALSE)

head(xyt_data)

}
\references{
Kieslich, P. J., Henninger, F., Wulff, D. U., Haslbeck, J. M. B., & Schulte-Mecklenbeck, M. (2019). Mouse-tracking: A practical guide to
implementation and analysis. In M. Schulte-Mecklenbeck, A. Kühberger, & J. G. Johnson (Eds.), A Handbook of Process Tracing
Methods (pp. 111-130). New York, NY: Routledge.
}
\author{
D. Schmitz
}
