% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_y.R
\name{extract_y}
\alias{extract_y}
\title{Extract y coordinates of time-normalized mouse-tracking data}
\usage{
extract_x(
  tn_data,
  ID_column,
  timestamps)
}
\arguments{
\item{tn_data}{The time-normalized mouse-tracking data. Can be extracted from the \code{mousetrap} object via \code{df$tn_trajectories}.}

\item{ID_column}{To maintain identification of individual trials, specify your ID variable. It must be part of \code{tn_data}. This can be achieved via \code{tn_data$ID <- df$data$ID}.}

\item{timestamps}{Specify the number of timestamps you have used for time-normalizing.}
}
\value{
A data frame.
}
\description{
\code{extract_y} extracts the y coordinates from a time-normalized \code{mousetrap} object.
}
\examples{
# extract time-normalized data
tn_tracks <- as.data.frame(mt_data$tn_trajectories)

# add ID columns
tn_tracks$ID <- mt_data$data$ID

# use function
extract_y(tn_data = tn_tracks,
ID_column = tn_tracks$ID,
timestamps = 138)

}
\references{
Kieslich, P. J., Henninger, F., Wulff, D. U., Haslbeck, J. M. B., & Schulte-Mecklenbeck, M. (2019). Mouse-tracking: A practical guide to
implementation and analysis. In M. Schulte-Mecklenbeck, A. Kühberger, & J. G. Johnson (Eds.), A Handbook of Process Tracing
Methods (pp. 111-130). New York, NY: Routledge.
}
\author{
D. Schmitz
}
