% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_parametric_plot.R
\name{facet_parametric_plot}
\alias{facet_parametric_plot}
\title{Plot a multi-panel plot of a parametric predictor across multiple QGAMs in a MQGAM}
\usage{
facet_parametric_plot(
  qgam,
  pred,
  cond = NULL,
  print.summary = FALSE,
  order = NULL,
  ncol = 1,
  xlab = "fit",
  ylab = NULL,
  scales = "free",
  size = 0.5,
  color = NULL,
  alpha = 1)
}
\arguments{
\item{qgam}{An mqgam object created with \code{qgam::mqgam}.}

\item{pred}{The predictor term to plot. Note: This is no longer identical to the \code{pred} argument specified for \code{itsadug::plot_parametric}.}

\item{cond}{A named list of the values to use for the other predictor terms (not in view). Used for choosing between smooths that share the same view predictors.}

\item{print.summary}{Logical: whether or not to print summary.}

\item{order}{Specify the order with which the levels given in \code{pred} should be plotted.}

\item{ncol}{The number of columns of the multi-panel plot.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{scales}{Should scales be free (\code{"free"}, the default) or fixed (\code{"fixed"})?}

\item{size}{Size argument for the ggplot object; specifies the size of points and lines.}

\item{color}{Color argument for the ggplot object; specifies the color of points and lines.}

\item{alpha}{Alpha argument for the ggplot object; specifies the transparency of points and lines.}
}
\value{
A ggplot object.
}
\description{
\code{facet_parametric_plot} creates a multi-panel plot of point plots with confidence interval ranges using \code{ggplot2}.
}
\examples{

# basic example
facet_parametric_plot(qgam = mtqgam_mqgam,
  pred = "factor_3")

# combining facet_parametric_plot with ggplot2
facet_parametric_plot(qgam = mtqgam_mqgam,
  pred = "factor_3") +
  theme_void() +
  labs(subtitle = "This is a subtitle")

}
\references{
Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast calibrated additive quantile regression. URL: https://arxiv.org/abs/1707.03307

van Rij J, Wieling M, Baayen R, & van Rijn H (2020). itsadug: Interpreting Time Series and Autocorrelated Data Using GAMMs. R package version 2.4.

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York.
}
\author{
D. Schmitz
}
