% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_smooth_plot.R
\name{facet_smooth_plot}
\alias{facet_smooth_plot}
\title{Plot a multi-panel plot of a smooth predictor across multiple QGAMs in a MQGAM}
\usage{
better_smooth_plot(
  qgam,
  pred,
  ncol = 1,
  xlab = NULL,
  ylab = "fit",
  scales = "free",
  size = 0.5,
  fill = "steelblue2",
  color = "black",
  alpha = 1)
}
\arguments{
\item{qgam}{An mqgam object created with \code{qgam::mqgam}.}

\item{pred}{The predictor term to plot. Note: This is no longer identical to the \code{pred} argument specified for \code{itsadug::plot_parametric}.}

\item{ncol}{The number of columns of the multi-panel plot.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{scales}{Should scales be free (\code{"free"}, the default) or fixed (\code{"fixed"})?}

\item{size}{Size argument for the ggplot object; specifies the size of the line.}

\item{fill}{Color argument for the ggplot object; specifies the color of the confidence interval.}

\item{color}{Color argument for the ggplot object; specifies the color of the line.}

\item{alpha}{Alpha argument for the ggplot object; specifies the transparency of the confidence interval.}
}
\value{
A ggplot object.
}
\description{
\code{facet_smooth_plot} creates a multi-panel plot of smooth term plots with confidence intervals using \code{ggplot2}.
}
\examples{

# basic example
facet_smooth_plot(qgam = mqgams_x,
  pred = "order")

# combining facet_smooth_plot with ggplot2
facet_smooth_plot(qgam = mqgams_x,
  pred = "order") +
  theme_void() +
  labs(subtitle = "This is a subtitle")

}
\references{
Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast calibrated additive quantile regression. URL: https://arxiv.org/abs/1707.03307

van Rij J, Wieling M, Baayen R, & van Rijn H (2020). itsadug: Interpreting Time Series and Autocorrelated Data Using GAMMs. R package version 2.4.

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York.
}
\author{
D. Schmitz
}
