% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtqgam_mousetrap.R
\docType{data}
\name{mtqgam_mousetrap}
\alias{mtqgam_mousetrap}
\title{Example mousetrap object}
\format{
A \code{mousetrap} object with 3 elements, i.e. data, mouse-tracks, and time-normalized mouse-tracks.
}
\source{
Schmitz, Dominic (2021). mtqgam: Mouse-Tracking Data in QGAMs. \url{https://github.com/dosc91/mtqgam}
}
\usage{
mtqgam_mousetrap
}
\description{
A \code{mousetrap} object.
}
\keyword{datasets}
