% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtqgam_mqgam.R
\docType{data}
\name{mtqgam_mqgam}
\alias{mtqgam_mqgam}
\title{Example MQGAM object}
\format{
An \code{mqgam} object with 5 elements, i.e. fitted to 5 quantiles.
}
\source{
Schmitz, Dominic (2021). mtqgam: Mouse-Tracking Data in QGAMs. \url{https://github.com/dosc91/mtqgam}
}
\usage{
mtqgam_mqgam
}
\description{
An \code{mqgam} object created with \code{qgam::mqgam}.
}
\keyword{datasets}
