% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtqgam_qgam.R
\docType{data}
\name{mtqgam_qgam}
\alias{mtqgam_qgam}
\title{Example QGAM object}
\format{
An \code{qgam} object with 58 elements.
}
\source{
Schmitz, Dominic (2021). mtqgam: Mouse-Tracking Data in QGAMs. \url{https://github.com/dosc91/mtqgam}
}
\usage{
mtqgam_qgam
}
\description{
A \code{qgam} object created with \code{qgam::qgam}.
}
\keyword{datasets}
