% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtqgam_raw_data.R
\docType{data}
\name{mtqgam_raw_data}
\alias{mtqgam_raw_data}
\title{Example data to create QGAMs with}
\format{
A data frame with 5000 rows and 6 variables:
\describe{
\item{dependent}{a numeric variable of random values; used as dependent variable}
\item{numeric_1}{a numeric variable of random values}
\item{numeric_2}{a numeric variable of random values}
\item{factor_1}{a factor variable with 4 levels}
\item{factor_2}{a factor variable with 5 levels}
\item{factor_3}{a factor variable with 2 levels}
}
}
\source{
Schmitz, Dominic (2021). mtqgam: Mouse-Tracking Data in QGAMs. \url{https://github.com/dosc91/mtqgam}
}
\usage{
mtqgam_raw_data
}
\description{
A dataset containing simulated data to create \code{qgam::qgam} and \code{qgam::mqgam} objects with.
}
\keyword{datasets}
