% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparison.R
\name{pairwise_comparison}
\alias{pairwise_comparison}
\title{Pairwise Tukey Comparisons}
\usage{
pairwise_comparison(
  qgam,
  quantile = NULL,
  factor, by = NULL,
  ...)
}
\arguments{
\item{qgam}{A qgam object created with \code{qgam::qgam} or extracted from a \code{qgam::mqgam} object, or a collection of qgams created with \code{qgam::mqgam}.}

\item{quantile}{If \code{qgam} is a collection of qgam models, specify the quantile you are interested in. Not meaningful for single qgam objects.}

\item{factor}{The categorical variable you wish to compute Tukey comparisons for. Must be one of the predictor or smooth terms given in the qgam formula.}

\item{...}{Any of the other arguments \code{emmeans::emmeans} can take.}
}
\description{
\code{pairwise_comparison} computes pairwise Tukey comparisons for any combination of variables in a given QGAM using
the \code{emmeans} package. Pairwise Tukey comparisons are most sensible to check differences between levels of a
categorical predictor.
}
\examples{

pairwise_comparison(qgam = mtqgam_mqgam, quantile = 0.5, "factor_1")

}
\references{
Russell V. Lenth (2021). emmeans: Estimated Marginal Means, aka Least-Squares Means. R package version 1.5.4. https://CRAN.R-project.org/package=emmeans
}
\author{
D. Schmitz
}
