% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqgam.R
\name{summary.mqgam}
\alias{summary.mqgam}
\title{Summary for a MQGAM fit}
\usage{
## S3 method for class 'mqgam'
summary(mqgam)
}
\arguments{
\item{mqgam}{An mqgam object created with \code{qgam::mqgam}.}
}
\value{
Please see \link[mgcv]{summary.gam} for a detailed explanation of the values given.
}
\description{
Takes a fitted \code{mqgam} object produced by \code{qgam::mqgam} and produces various useful summaries from it, making use of
the \code{mgcv::summary.gam} method.
}
\details{
Please see \link[mgcv]{summary.gam} for a detailed explanation of the summaries given.
}
\examples{
# general usage
summary(mqgam = mtqgam_mqgam)

# printing just one quantile summary
summary(mqgam = mtqgam_mqgam)[[3]]

}
\references{
Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast calibrated additive quantile regression. URL: https://arxiv.org/abs/1707.03307

Wood, S.N. (2011). Fast stable restricted maximum likelihood and marginal likelihood estimation of semiparametric generalized linear models. Journal of the Royal Statistical Society (B) 73(1), 3-36.
}
\author{
D. Schmitz
}
