% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_interaction_plot.R
\name{xy_interaction_plot}
\alias{xy_interaction_plot}
\title{Plot XY plots of parametric predictor interactions in QGAMs}
\usage{
xy_interaction_plot(
  qgam_x,
  qgam_y,
  quantile = NULL,
  pred,
  cond = NULL,
  print.summary = FALSE,
  order = NULL,
  ncol = 2,
  xlab = "X coordinates",
  ylab = "Y coordinates",
  scales = "free",
  size = 3,
  color = NULL,
  alpha = 1)
}
\arguments{
\item{qgam_x}{The estimates plotted on the X axis. A qgam object created with \code{qgam::qgam} or extracted from a \code{qgam::mqgam} object, or a collection of qgams created with \code{qgam::mqgam}.}

\item{qgam_y}{The estimates plotted on the Y axis. A qgam object created with \code{qgam::qgam} or extracted from a \code{qgam::mqgam} object, or a collection of qgams created with \code{qgam::mqgam}.}

\item{quantile}{If \code{qgam} is a collection of qgam models, specify the quantile you are interested in. Not meaningful for single qgam objects.}

\item{pred}{The predictor term to plot. Note: This is no longer identical to the \code{pred} argument specified for \code{itsadug::plot_parametric}.}

\item{cond}{A named list of the values to use for the other predictor terms (not in view). Used for choosing between smooths that share the same view predictors.}

\item{print.summary}{Logical: whether or not to print summary.}

\item{order}{Specify the order with which the levels given in \code{pred} should be plotted.}

\item{ncol}{The number of columns of the multi-panel plot.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{scales}{Should scales be free (\code{"free"}, the default) or fixed (\code{"fixed"})?}

\item{size}{Size argument for the ggplot object; specifies the size of points and lines.}

\item{color}{Color argument for the ggplot object; specifies the color of points and lines.}

\item{alpha}{Alpha argument for the ggplot object; specifies the transparency of points and lines.}
}
\value{
A ggplot object.
}
\description{
\code{xy_interaction_plot} creates a point plot with confidence interval ranges using \code{ggplot2}. It combines the model estimates for the same predictor
in two (sets of) QGAMs, i.e. it displays estimates of X and Y coordinates simultaneously in a combined plot.
It makes use of \code{mtqgam}'s \code{better_parametric_plot} and thus outputs the same information in your console.
}
\examples{

# using a single qgam extracted from an mqgam object OR fitted with qgam::qgam
xy_interaction_plot(qgam_x = mtqgam_qgam,
qgam_y = mtqgam_qgam,
pred = "factor_2",
cond = "factor_3")

# using a qgam that is part of an mqgam object
xy_interaction_plot(qgam_x = mtqgam_mqgam,
qgam_y = mtqgam_mqgam,
pred = "factor_2",
cond = "factor_3")

# specifying color
xy_interaction_plot(qgam_x = mtqgam_mqgam,
qgam_y = mtqgam_mqgam,
pred = "factor_2",
cond = "factor_3",
color = c("blue", "purple", "green", "yellow", "pink"))

# combining better_interaction_plot with ggplot2
xy_interaction_plot(qgam_x = mtqgam_mqgam,
qgam_y = mtqgam_mqgam,
pred = "factor_2",
cond = "factor_3") +
theme_void() +
labs(subtitle = "This is a subtitle")

}
\references{
Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast calibrated additive quantile regression. URL: https://arxiv.org/abs/1707.03307

van Rij J, Wieling M, Baayen R, & van Rijn H (2020). itsadug: Interpreting Time Series and Autocorrelated Data Using GAMMs. R package version 2.4.

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York.
}
\author{
D. Schmitz
}
