% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_parametric_plot.R
\name{better_parametric_plot}
\alias{better_parametric_plot}
\title{Plot better plots of parametric predictors in QGAMs}
\usage{
better_parametric_plot(
  qgam,
  quantile = NULL,
  pred,
  cond = NULL,
  print.summary = FALSE,
  plot.old = FALSE,
  order = NULL,
  xlab = "fit",
  ylab = NULL,
  size = 0.5,
  color = NULL,
  alpha = 1)
}
\arguments{
\item{qgam}{A qgam object created with \code{qgam::qgam} or extracted from a \code{qgam::mqgam} object, or a collection of qgams created with \code{qgam::mqgam}.}

\item{quantile}{If \code{qgam} is a collection of qgam models, specify the quantile you are interested in. Not meaningful for single qgam objects.}

\item{pred}{The predictor term to plot. Note: This is no longer identical to the \code{pred} argument specified for \code{itsadug::plot_parametric}.}

\item{cond}{A named list of the values to use for the other predictor terms (not in view). Used for choosing between smooths that share the same view predictors.}

\item{print.summary}{Logical: whether or not to print summary.}

\item{plot.old}{Plot the original \code{plot_parametric} as well.}

\item{order}{Specify the order with which the levels given in \code{pred} should be plotted.}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{size}{Size argument for the ggplot object; specifies the size of points and lines.}

\item{color}{Color argument for the ggplot object; specifies the color of points and lines.}

\item{alpha}{Alpha argument for the ggplot object; specifies the transparency of points and lines.}
}
\value{
A ggplot object.
}
\description{
\code{better_parametric_plot} creates a point plot with confidence interval ranges using \code{ggplot2}. It basically is a wrapper for \code{itsadug}'s
\code{plot_parametric} function.
}
\examples{

# using a single qgam extracted from an mqgam object OR fitted with qgam::qgam
better_parametric_plot(qgam = mtqgam_qgam,
  pred = "factor_3")

# using a qgam that is part of an mqgam object
library(qgam)

better_parametric_plot(qgam = mtqgam_mqgam,
  quantile = 0.5,
  pred = "factor_3")

# combining better_parametric_plot with ggplot2
better_parametric_plot(qgam = mtqgam_qgam,
  pred = "factor_3") +
  theme_void() +
  labs(subtitle = "This is a subtitle")

}
\references{
Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast calibrated additive quantile regression. URL: https://arxiv.org/abs/1707.03307

van Rij J, Wieling M, Baayen R, & van Rijn H (2020). itsadug: Interpreting Time Series and Autocorrelated Data Using GAMMs. R package version 2.4.

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York.
}
\author{
D. Schmitz
}
