% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_smooth_plot.R
\name{better_smooth_plot}
\alias{better_smooth_plot}
\title{Plot better plots of smooth predictors in QGAMs}
\usage{
better_parametric_plot(
  qgam,
  smooth_term,
  size,
  fill,
  alpha,
  color)
}
\arguments{
\item{qgam}{A qgam object, created with \code{qgam} or extracted from a \code{mqgam} object.}

\item{smooth_term}{The smooth term you wish to plot. Must be one of the smooth terms given in the qgam formula.}

\item{plot.old}{Plot the original \code{plot} as well.}

\item{size}{Size argument for the ggplot object; specifies the size of the line.}

\item{fill}{Color argument for the ggplot object; specifies the color of the confidence interval.}

\item{alpha}{Alpha argument for the ggplot object; specifies the transparency of the confidence interval.}

\item{color}{Color argument for the ggplot object; specifies the color of the line.}
}
\value{
A ggplot object.
}
\description{
\code{better_smooth_plot} creates a smooth plot with confidence intervals using \code{ggplot2}. It basically is a wrapper for the base \code{plot} function.
}
\examples{
better_smooth_plot(qgam = tmp.x.3,
  smooth_term = "Age")

better_smooth_plot(qgam = tmp.x.3,
  smooth_term = "Age",
  size = 1,
  fill = "purple",
  alpha = 0.5,
  color = "darkgreen") +
  labs(subtitle = "You can basically add all ggplot functions and arguments you are familiar with.")

}
\references{
Fasiolo M., Goude Y., Nedellec R., & Wood S. N. (2017). Fast calibrated additive quantile regression. URL: https://arxiv.org/abs/1707.03307

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York.
}
\author{
D. Schmitz
}
